Created Mocks subfolder and moved some classes there
