+++ Aegis.Pumps.Tests/Aegis.Pumps.Tests.csproj
@@ -73,8 +73,8 @@
     <Compile Include="BlackListClientTests.cs" />
     <Compile Include="ClientTests.cs" />
     <Compile Include="HTTPRequestMetadataCacheTests.cs" />
-    <Compile Include="MockAegisServiceClient.cs" />
-    <Compile Include="MockHTTPRequestMetadata.cs" />
+    <Compile Include="Mocks\MockAegisServiceClient.cs" />
+    <Compile Include="Mocks\MockHTTPRequestMetadata.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SettingsOnlineDataTests.cs" />
   </ItemGroup>
