+++ Aegis.Pumps.Tests/HTTPRequestMetadataCacheTests.cs
@@ -679,7 +679,7 @@ using System;
 using System.Collections.Concurrent;
 using System.Net;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Phobos.Tests;
+using Aegis.Pumps.Tests.Mocks;
 
 namespace Aegis.Pumps.Tests
 {
